refactor: extract MovieList component
